Add admin users
This patch adds an `admin` flag to each users
When this flag is set and said user is logged in a new button appears
on the navigation bar. The button redirects the user to the `/admin`
endpoint, where they will be able to do management tasks.

For now the admin panel is just a simple lorem ipsum with a few headers
but in the future it will have more features such as content moderation,
city creation and approval as well as user management.